move work-status-renderer component to lib
