+++ src-ts/lib/index.ts
@@ -22,3 +22,4 @@ export * from './svgs'
 export * from './table'
 export * from './tabs-navbar'
 export * from './work-provider'
+export * from './work-status-renderer'
