+++ src-ts/lib/work-status-renderer/WorkStatusRenderer.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../lib/styles';
+@import '../styles';
 
 .status-container {
     display: flex;
