+++ src-ts/tools/work/work-table/work-table.config.tsx
@@ -1,8 +1,7 @@
-import { TableColumn, Work } from '../../../lib'
+import { TableColumn, Work, WorkStatusRenderer } from '../../../lib'
 
 import { WorkBadgeRenderer } from './work-badge-renderer'
 import { WorkDeleteButtonRenderer } from './work-delete-button-renderer'
-import { WorkStatusRenderer } from './work-status-renderer'
 import { WorkTableTitleRenderer } from './work-table-title-renderer'
 
 function messageBadgeRenderer(work: Work): JSX.Element {
